Helpify-AI-main/project/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
scikit-learn>=1.3.0
pandas>=2.0.3
numpy>=1.24.3
matplotlib>=3.7.2
seaborn>=0.12.2
pyyaml>=6.0
jupyter>=1.0.0
ipython>=8.14.0
requests>=2.31.0
tqdm>=4.65.0
click>=8.1.6
python-dotenv>=1.0.0
nltk>=3.8.1
tokenizers>=0.13.3
transformers>=4.30.2
datasets>=2.13.0
tensorboard>=2.13.0
flask>=2.3.2
gunicorn>=20.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.1
